# Conflicts:
#	Telegram/CMakeLists.txt
#	Telegram/Resources/uwp/AppX/AppxManifest.xml
#	Telegram/Resources/winrc/Telegram.rc
#	Telegram/Resources/winrc/Updater.rc
#	Telegram/SourceFiles/boxes/about_box.cpp
#	Telegram/SourceFiles/boxes/about_sponsored_box.cpp
#	Telegram/SourceFiles/boxes/background_preview_box.cpp
#	Telegram/SourceFiles/boxes/local_storage_box.cpp
#	Telegram/SourceFiles/boxes/peers/edit_peer_color_box.cpp
#	Telegram/SourceFiles/boxes/select_future_owner_box.cpp
#	Telegram/SourceFiles/boxes/star_gift_auction_box.cpp
#	Telegram/SourceFiles/boxes/star_gift_preview_box.cpp
#	Telegram/SourceFiles/boxes/url_auth_box_content.cpp
#	Telegram/SourceFiles/calls/group/calls_group_rtmp.cpp
#	Telegram/SourceFiles/chat_helpers/ttl_media_layer_widget.cpp
#	Telegram/SourceFiles/core/version.h
#	Telegram/SourceFiles/data/data_types.h
#	Telegram/SourceFiles/dialogs/dialogs_inner_widget.cpp
#	Telegram/SourceFiles/dialogs/ui/dialogs_top_bar_suggestion_content.cpp
#	Telegram/SourceFiles/dialogs/ui/posts_search_intro.cpp
#	Telegram/SourceFiles/export/view/export_view_progress.cpp
#	Telegram/SourceFiles/export/view/export_view_settings.cpp
#	Telegram/SourceFiles/history/history_item.cpp
#	Telegram/SourceFiles/history/history_widget.cpp
#	Telegram/SourceFiles/history/view/controls/history_view_compose_controls.cpp
#	Telegram/SourceFiles/history/view/history_view_contact_status.cpp
#	Telegram/SourceFiles/history/view/history_view_pinned_bar.cpp
#	Telegram/SourceFiles/history/view/history_view_scheduled_section.cpp
#	Telegram/SourceFiles/history/view/history_view_sticker_toast.cpp
#	Telegram/SourceFiles/history/view/history_view_translate_bar.cpp
#	Telegram/SourceFiles/info/channel_statistics/boosts/create_giveaway_box.cpp
#	Telegram/SourceFiles/info/channel_statistics/boosts/info_boosts_inner_widget.cpp
#	Telegram/SourceFiles/info/channel_statistics/earn/info_channel_earn_list.cpp
#	Telegram/SourceFiles/info/info.style
#	Telegram/SourceFiles/info/peer_gifts/info_peer_gifts_widget.cpp
#	Telegram/SourceFiles/info/polls/info_polls_list_widget.cpp
#	Telegram/SourceFiles/info/profile/info_profile_actions.cpp
#	Telegram/SourceFiles/info/profile/info_profile_top_bar.cpp
#	Telegram/SourceFiles/info/stories/info_stories_inner_widget.cpp
#	Telegram/SourceFiles/info/stories/info_stories_widget.cpp
#	Telegram/SourceFiles/info/userpic/info_userpic_emoji_builder.cpp
#	Telegram/SourceFiles/inline_bots/inline_results_inner.cpp
#	Telegram/SourceFiles/intro/intro_widget.cpp
#	Telegram/SourceFiles/iv/iv_controller.cpp
#	Telegram/SourceFiles/media/stories/media_stories_controller.cpp
#	Telegram/SourceFiles/media/view/media_view_overlay_widget.cpp
#	Telegram/SourceFiles/media/view/media_view_playback_sponsored.cpp
#	Telegram/SourceFiles/passport/passport_panel_edit_contact.cpp
#	Telegram/SourceFiles/passport/passport_panel_password.cpp
#	Telegram/SourceFiles/payments/ui/payments_form_summary.cpp
#	Telegram/SourceFiles/settings/cloud_password/settings_cloud_password_common.cpp
#	Telegram/SourceFiles/settings/sections/settings_credits.cpp
#	Telegram/SourceFiles/settings/sections/settings_folders.cpp
#	Telegram/SourceFiles/settings/sections/settings_local_passcode.cpp
#	Telegram/SourceFiles/settings/sections/settings_main.cpp
#	Telegram/SourceFiles/settings/sections/settings_passkeys.cpp
#	Telegram/SourceFiles/settings/settings_credits_graphics.cpp
#	Telegram/SourceFiles/statistics/chart_widget.cpp
#	Telegram/SourceFiles/ui/boxes/collectible_info_box.cpp
#	Telegram/SourceFiles/ui/boxes/confirm_phone_box.cpp
#	Telegram/SourceFiles/ui/boxes/show_or_premium_box.cpp
#	Telegram/SourceFiles/ui/chat/choose_theme_controller.cpp
#	Telegram/SourceFiles/ui/chat/group_call_bar.cpp
#	Telegram/SourceFiles/ui/controls/invite_link_buttons.cpp
#	Telegram/SourceFiles/ui/controls/location_picker.cpp
#	Telegram/SourceFiles/ui/controls/table_rows.cpp
#	Telegram/SourceFiles/window/themes/window_theme_warning.cpp
#	Telegram/SourceFiles/window/window_lock_widgets.cpp
#	Telegram/SourceFiles/window/window_peer_menu.cpp
#	Telegram/SourceFiles/window/window_setup_email.cpp
#	Telegram/build/version
#	Telegram/lib_webview

+++ Telegram/SourceFiles/ffmpeg/ffmpeg_utility.cpp
@@ -691,7 +691,7 @@ QImage CreateFrameStorage(QSize size) {
 		? (widthAlign - (width % widthAlign))
 		: 0);
 	const auto perLine = neededWidth * kPixelBytesSize;
-	const auto buffer = new uchar[perLine * height + kAlignImageBy];
+	const auto buffer = new uchar[size_t(perLine) * height + kAlignImageBy];
 	const auto cleanupData = static_cast<void *>(buffer);
 	const auto address = reinterpret_cast<uintptr_t>(buffer);
 	const auto alignedBuffer = buffer + ((address % kAlignImageBy)
